Revised summary.cross so that it pays attention to options("width") and prints things more nicely if there are loads and loads of phenotypes or whatever.
The title of the plot produced by plot.rf can now be modified by
including the argument 'main' in the call.

In plot.geno, if chr is missing, we plot the genotypes for the first
chromosome. 

In geno.crosstab, changed the column and row labels for missing data
from "NA" to "-".


git-svn-id: file:///home/wrk/tmp/qtl@52 233cfdfd-cf43-0410-8e6e-a5ad47e32483