Fixed a bug in stepwiseqtl, which involved fixes in parseQTLformula and countqtlterms.
In refineqtl, the default for the argument keeplodprofile is now
TRUE.  The LOD profiles returned by refineqtl are now of class
"scanone".

lodint and bayesint were revised to accept qtl objects output by
refineqtl (with keeplodprofile=TRUE).  An additional argument,
qtl.index, was added to indicate for which QTL (within such qtl
objects) the approximate confidence intervals should be derived.
For scanone output, the functions were modified so that, if the
results concern just a single chromosome, the chromosome argument is
not needed.


git-svn-id: file:///home/wrk/tmp/qtl@58 233cfdfd-cf43-0410-8e6e-a5ad47e32483