Fixed a bug in matchchr (used for subsetting chromosomes).
Revised the way in which LOD score columns are renamed in c.scanone
and cbind.scanoneperm. If labels are given, these are appended to
the end of the names, but if labels are not given and there are no
repeats in the column names, the column names are left as they were.


git-svn-id: file:///home/wrk/tmp/qtl@97 233cfdfd-cf43-0410-8e6e-a5ad47e32483