permutations on QTL effect on trait variance
scanonevar() models two effects of the QTL: (1) the mean effect and (2) the variance effect.  This function runs permutations where the QTL in the variance term gets permuted but the QTL in the mean term does not.  This is the valid permutation test (technically a randomization test) for the QTL effect on variance.

Former-commit-id: 6e596243bfe45cc960dc4440927045bbf4e8cc22