changed scanonevar.meanperm() such that additive and dominance design matrices receive the same permutation of the genetic variant in question.
Created scanonevar.varperm(), very similarly to scanonevar.meanperm().

Created scanonevar.meanvarperm(), which uses one permutation of the design matrix for the additive and dominance effects on the mean, and another permutation for the additive and dominance effects on the variance.

Created scanonevar.phenperm(), which permutes the phenotypes.

Creating all of these is a part of the process of thinking through which of them is most appropriate.