Created is.cross.R
This file has two functions: 
is.cross, a utility function for testing whether an object is a cross object
is.f2.cross, a utility function for testing whether an object is an F2 cross object

Not sure if this functionality is somewhere else in the package, but I couldn't find it.  Of course, these functions could be extended to other types of crosses, but this is where I've started.